+++ osu.Game.Tests/Visual/Online/TestSceneChatOverlay.cs
@@ -96,6 +96,16 @@ namespace osu.Game.Tests.Visual.Online
             AddAssert("Selector is visible", () => chatOverlay.SelectionOverlayState == Visibility.Visible);
         }
 
+        [Test]
+        public void TestLeaveChannelThenJoiningNew()
+        {
+            AddStep("Join channel 1", () => channelManager.JoinChannel(channel1));
+            AddStep("Close channel 1", () => clickDrawable(((TestChannelTabItem)chatOverlay.TabMap[channel1]).CloseButton.Child));
+            AddAssert("Current channel is null", () => channelManager.CurrentChannel.Value == null);
+            AddStep("Join channel 1", () => channelManager.JoinChannel(channel1));
+            AddAssert("Current channel is channel 1", () => channelManager.CurrentChannel.Value == channel1);
+        }
+
         private void clickDrawable(Drawable d)
         {
             InputManager.MoveMouseTo(d);
